requirements.txt
asgiref==3.10.0
dj-rest-auth==7.0.1
Django==5.2.7
django-allauth==65.12.1
django-cors-headers==4.9.0
django-extensions==4.1
django-environ==0.10.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-nested-routers==0.95.0
drf-yasg==1.21.11
inflection==0.5.1
packaging==25.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.3
sqlparse==0.5.3
uritemplate==4.2.0
Pillow==12.0.0
pytest
requests
pytest-django